Python/Django/django-basics/pyproject.toml
[tool.poetry]
name = "django-basics"
version = "0.1.0"
description = ""
authors = ["Youngjun Lee <jun2korea@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
djangorestframework = "^3.14.0"
django-silk = "^5.1.0"
django-stubs = "^4.2.7"
djangorestframework-stubs = "^3.14.5"
mypy = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
